// src/modules/permissions/enums/permission.enum.ts
export enum ENUM_PERMISSION_SUBJECT {
  ADMIN_DASHBOARD = "ADMIN_DASHBOARD",
  ADMIN_STAFF = "ADMIN_STAFF",
  ADMIN_USER = "ADMIN_USER",
  ADMIN_GROUP_ROLE = "ADMIN_GROUP_ROLE",
  ADMIN_ROLE = "ADMIN_ROLE",
  ADMIN_CATEGORY = "ADMIN_CATEGORY",
  ADMIN_VARIANT_TYPE = "ADMIN_VARIANT_TYPE",
  ADMIN_VARIANT_TYPE_VALUE = "ADMIN_VARIANT_TYPE_VALUE",
  ADMIN_PRODUCT = "ADMIN_PRODUCT",
  ADMIN_VARIANT_PRODUCT = "ADMIN_VARIANT_PRODUCT",
  ADMIN_PRODUCT_INVENTORY = "ADMIN_PRODUCT_INVENTORY",
  ADMIN_VARIANT_PRODUCT_INVENTORY = "ADMIN_VARIANT_PRODUCT_INVENTORY",
  ADMIN_CART = "ADMIN_CART",
}

export enum ENUM_PERMISSION {
  ADMIN_DASHBOARD_VIEW = "ADMIN_DASHBOARD.VIEW",
  ADMIN_DASHBOARD_CREATE = "ADMIN_DASHBOARD.CREATE",
  ADMIN_DASHBOARD_UPDATE = "ADMIN_DASHBOARD.UPDATE",
  ADMIN_DASHBOARD_DELETE = "ADMIN_DASHBOARD.DELETE",

  // STAFF
  ADMIN_STAFF_VIEW = "ADMIN_STAFF.VIEW",
  ADMIN_STAFF_CREATE = "ADMIN_STAFF.CREATE",
  ADMIN_STAFF_UPDATE = "ADMIN_STAFF.UPDATE",
  ADMIN_STAFF_DELETE = "ADMIN_STAFF.DELETE",

  // User
  ADMIN_USER_VIEW = "ADMIN_USER.VIEW",
  ADMIN_USER_CREATE = "ADMIN_USER.CREATE",
  ADMIN_USER_UPDATE = "ADMIN_USER.UPDATE",
  ADMIN_USER_DELETE = "ADMIN_USER.DELETE",

  // Group Role
  ADMIN_GROUP_ROLE_VIEW = "ADMIN_GROUP_ROLE.VIEW",
  ADMIN_GROUP_ROLE_CREATE = "ADMIN_GROUP_ROLE.CREATE",
  ADMIN_GROUP_ROLE_UPDATE = "ADMIN_GROUP_ROLE.UPDATE",
  ADMIN_GROUP_ROLE_DELETE = "ADMIN_GROUP_ROLE.DELETE",

  // Role
  ADMIN_ROLE_VIEW = "ADMIN_ROLE.VIEW",
  ADMIN_ROLE_CREATE = "ADMIN_ROLE.CREATE",
  ADMIN_ROLE_UPDATE = "ADMIN_ROLE.UPDATE",
  ADMIN_ROLE_DELETE = "ADMIN_ROLE.DELETE",

  // Category
  ADMIN_CATEGORY_VIEW = "ADMIN_CATEGORY.VIEW",
  ADMIN_CATEGORY_CREATE = "ADMIN_CATEGORY.CREATE",
  ADMIN_CATEGORY_UPDATE = "ADMIN_CATEGORY.UPDATE",
  ADMIN_CATEGORY_DELETE = "ADMIN_CATEGORY.DELETE",

  // Variant Type
  ADMIN_VARIANT_TYPE_VIEW = "ADMIN_VARIANT_TYPE.VIEW",
  ADMIN_VARIANT_TYPE_CREATE = "ADMIN_VARIANT_TYPE.CREATE",
  ADMIN_VARIANT_TYPE_UPDATE = "ADMIN_VARIANT_TYPE.UPDATE",
  ADMIN_VARIANT_TYPE_DELETE = "ADMIN_VARIANT_TYPE.DELETE",

  // Variant Type Value
  ADMIN_VARIANT_TYPE_VALUE_VIEW = "ADMIN_VARIANT_TYPE_VALUE.VIEW",
  ADMIN_VARIANT_TYPE_VALUE_CREATE = "ADMIN_VARIANT_TYPE_VALUE.CREATE",
  ADMIN_VARIANT_TYPE_VALUE_UPDATE = "ADMIN_VARIANT_TYPE_VALUE.UPDATE",
  ADMIN_VARIANT_TYPE_VALUE_DELETE = "ADMIN_VARIANT_TYPE_VALUE.DELETE",

  // Product
  ADMIN_PRODUCT_VIEW = "ADMIN_PRODUCT.VIEW",
  ADMIN_PRODUCT_CREATE = "ADMIN_PRODUCT.CREATE",
  ADMIN_PRODUCT_UPDATE = "ADMIN_PRODUCT.UPDATE",
  ADMIN_PRODUCT_DELETE = "ADMIN_PRODUCT.DELETE",

  // Variant Product
  ADMIN_VARIANT_PRODUCT_VIEW = "ADMIN_VARIANT_PRODUCT.VIEW",
  ADMIN_VARIANT_PRODUCT_CREATE = "ADMIN_VARIANT_PRODUCT.CREATE",
  ADMIN_VARIANT_PRODUCT_UPDATE = "ADMIN_VARIANT_PRODUCT.UPDATE",
  ADMIN_VARIANT_PRODUCT_DELETE = "ADMIN_VARIANT_PRODUCT.DELETE",

  // Product Inventory
  ADMIN_PRODUCT_INVENTORY_VIEW = "ADMIN_PRODUCT_INVENTORY.VIEW",
  ADMIN_PRODUCT_INVENTORY_CREATE = "ADMIN_PRODUCT_INVENTORY.CREATE",
  ADMIN_PRODUCT_INVENTORY_UPDATE = "ADMIN_PRODUCT_INVENTORY.UPDATE",
  ADMIN_PRODUCT_INVENTORY_DELETE = "ADMIN_PRODUCT_INVENTORY.DELETE",

  // Variant Product Inventory
  ADMIN_VARIANT_PRODUCT_INVENTORY_VIEW = "ADMIN_VARIANT_PRODUCT_INVENTORY.VIEW",
  ADMIN_VARIANT_PRODUCT_INVENTORY_CREATE = "ADMIN_VARIANT_PRODUCT_INVENTORY.CREATE",
  ADMIN_VARIANT_PRODUCT_INVENTORY_UPDATE = "ADMIN_VARIANT_PRODUCT_INVENTORY.UPDATE",
  ADMIN_VARIANT_PRODUCT_INVENTORY_DELETE = "ADMIN_VARIANT_PRODUCT_INVENTORY.DELETE",

  // Cart
  ADMIN_CART_VIEW = "ADMIN_CART.VIEW",
  ADMIN_CART_CREATE = "ADMIN_CART.CREATE",
  ADMIN_CART_UPDATE = "ADMIN_CART.UPDATE",
  ADMIN_CART_DELETE = "ADMIN_CART.DELETE",
}